Update: Search components tsx로 변경
